requirements.txt
langgraph>=0.2.0
langchain-anthropic>=0.3.0
langchain-community>=0.3.0
langchain-core>=0.3.0
tavily-python>=0.5.0
pdfplumber>=0.11.0
pandas>=2.2.0
python-dotenv>=1.0.0
